+++ CHANGELOG.md
@@ -18,6 +18,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - The log message listing custom formats without scores in the guide now prints information one per
   line (improved readability)
 - Duplicate custom formats in the guide now issue a warning and get skipped.
+- Do not invoke the Radarr API to update a quality profile if there are no updated scores inside it.
 
 ## [1.4.0] - 2021-05-14